Getting started
Began the test file and making slight modifications to the directory_oneroster.py master file. It is going to be a bastardized child of the directory_okta.py for a while now so as not to break wider functionality while tinkering with things and setting it up. ~Unclear if this is the best approach, but it's the one I'm using for now.