Sync Algebra1 users
This is a destructive, exploratory commit that should never be merged.

The directory_oneroster.py file, derived from directory_okta.py, handles the users from raw json data that pretends to be a listing of Algebra1 students. This is achieved through hardcoding Algebra1 into 'find_group' and then using a tweaked version of convert_user driven by simply observing the fields we have available in our dummy data file.

More work is required to get this to a point where it actually aligns with our strategy and is robust enough to handle when data isn't where we want it to be, but it's a start.